# app/endpoints/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
  title: REST API
  version: v1.0.0
paths:
  /vrf/v1/beacon:
    get:
      description: Beacon returns the latest VrfBeacon for the current or specified height
      operationId: v1_beacon
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryBeaconResponse'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
      summary: Beacon
      tags:
        - V1
  /vrf/v1/info:
    get:
      description: |-
        Info returns static drand chain information used for verification and
         debugging.
      operationId: v1_info
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryInfoResponse'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
      summary: Info
      tags:
        - V1
  /vrf/v1/params:
    get:
      description: Params returns the on-chain VRF parameters.
      operationId: v1_params
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryParamsResponse'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
      summary: Params
      tags:
        - V1
  /vrf/v1/random_words:
    get:
      description: RandomWords expands the current block's VRF seed into N random words
      operationId: v1_random_words
      parameters:
        - name: count
          in: query
          description: count is the number of random words to derive.
          schema:
            type: integer
            format: uint32
        - name: userSeed
          in: query
          description: |-
            user_seed is an application-supplied seed that is mixed into the
             derivation to give callers domain separation.
          schema:
            type: string
            format: bytes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryRandomWordsResponse'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
      summary: Random Words
      tags:
        - V1
  /vrf/v1/randomness:
    get:
      description: |-
        Randomness returns a drand beacon for a specific round or the latest
         available beacon when round is unset (zero).
      operationId: v1_randomness
      parameters:
        - name: round
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryRandomnessResponse'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
      summary: Randomness
      tags:
        - V1
components:
  schemas:
    GoogleProtobufAny:
      type: object
      properties:
        '@type':
          type: string
          description: The type of the serialized message.
      additionalProperties: true
      description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
    QueryBeaconResponse:
      type: object
      properties:
        beacon:
          $ref: '#/components/schemas/VrfBeacon'
      description: QueryBeaconResponse carries the canonical VrfBeacon for the requested height.
    QueryInfoResponse:
      type: object
      properties:
        chainHash:
          type: string
          description: chain_hash is the canonical identifier for the drand chain.
          format: bytes
        publicKey:
          type: string
          description: public_key is the collective BLS public key for the drand group.
          format: bytes
        periodSeconds:
          type: string
          description: period_seconds is the drand beacon period in seconds.
        genesisUnixSec:
          type: string
          description: genesis_unix_sec is the drand genesis time (UNIX seconds) for round 1.
      description: QueryInfoResponse carries static drand chain information.
    QueryParamsResponse:
      type: object
      properties:
        params:
          $ref: '#/components/schemas/VrfParams'
      description: |-
        QueryParamsResponse carries the current VRF parameters. The fields mirror
         x/vrf's VrfParams type; the module keeper converts between the two.
    QueryRandomWordsResponse:
      type: object
      properties:
        drandRound:
          type: string
        seed:
          type: string
          format: bytes
        words:
          type: array
          items:
            type: string
            format: bytes
      description: |-
        QueryRandomWordsResponse returns the underlying drand round, the raw seed,
         and the derived random words.
    QueryRandomnessResponse:
      type: object
      properties:
        drandRound:
          type: string
          description: drand_round is the drand round number.
        randomness:
          type: string
          description: randomness is SHA256(signature).
          format: bytes
        signature:
          type: string
          description: signature is the drand BLS signature for this round.
          format: bytes
        previousSignature:
          type: string
          description: |-
            previous_signature is the BLS signature of the previous round, required
             for chained-scheme verification.
          format: bytes
      description: QueryRandomnessResponse carries a single drand beacon.
    Status:
      type: object
      properties:
        code:
          type: integer
          description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
          format: int32
        message:
          type: string
          description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
        details:
          type: array
          items:
            $ref: '#/components/schemas/GoogleProtobufAny'
          description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
      description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
    VrfBeacon:
      type: object
      properties:
        drandRound:
          type: string
          description: drand_round is the drand round number.
        randomness:
          type: string
          description: randomness is SHA256(signature).
          format: bytes
        signature:
          type: string
          description: signature is the drand BLS signature for this round.
          format: bytes
        previousSignature:
          type: string
          description: |-
            previous_signature is the BLS signature of the previous round, required
             for chained-scheme verification.
          format: bytes
      description: |-
        VrfBeacon is the canonical drand beacon selected for a given block height.
         It carries only the fields required for full on-chain verification.
    VrfParams:
      type: object
      properties:
        chainHash:
          type: string
          description: chain_hash is the canonical identifier for the drand chain.
          format: bytes
        publicKey:
          type: string
          description: public_key is the collective BLS public key for the drand group.
          format: bytes
        periodSeconds:
          type: string
          description: period_seconds is the drand beacon period in seconds.
        genesisUnixSec:
          type: string
          description: genesis_unix_sec is the drand genesis time (UNIX seconds) for round 1.
        safetyMarginSeconds:
          type: string
          description: |-
            safety_margin_seconds is the safety margin applied when computing
             target_round(H) from block times.
        enabled:
          type: boolean
          description: enabled gates VRF behavior in ExtendVote/PreBlock.
        reshareEpoch:
          type: string
          description: reshare_epoch is the on-chain resharing signal / epoch counter.
        slashingGraceBlocks:
          type: string
          description: slashing_grace_blocks configures the slashing grace window.
      description: |-
        VrfParams mirrors the PRD definition and contains all cryptographic and timing
         context needed to verify drand beacons on-chain and map block time to drand
         rounds.
tags:
  - name: V1
    description: ""